Fix some circular imports + create torrentStore
