+++ src/components/Dashboard/TRC/RightClickMenu.vue
@@ -5,7 +5,7 @@ import MoveTorrentDialog from '@/components/Dialogs/MoveTorrentDialog.vue'
 import RenameTorrentDialog from '@/components/Dialogs/RenameTorrentDialog.vue'
 import ShareLimitDialog from '@/components/Dialogs/ShareLimitDialog.vue'
 import SpeedLimitDialog from '@/components/Dialogs/SpeedLimitDialog.vue'
-import { useDashboardStore, useDialogStore, useMaindataStore, usePreferenceStore } from '@/stores'
+import { useDashboardStore, useDialogStore, useMaindataStore, usePreferenceStore, useTorrentStore } from '@/stores'
 import { TRCMenuEntry } from '@/types/vuetorrent'
 import { computed } from 'vue'
 import { useI18n } from 'vue-i18n'
@@ -22,6 +22,7 @@ const dashboardStore = useDashboardStore()
 const dialogStore = useDialogStore()
 const maindataStore = useMaindataStore()
 const preferenceStore = usePreferenceStore()
+const torrentStore = useTorrentStore()
 
 const trcVisible = computed({
   get: () => props.modelValue,
@@ -31,20 +32,20 @@ const trcVisible = computed({
 const isMultiple = computed(() => dashboardStore.selectedTorrents.length > 1)
 const hashes = computed(() => dashboardStore.selectedTorrents)
 const hash = computed(() => hashes.value[0])
-const torrent = computed(() => maindataStore.getTorrentByHash(hash.value))
-const torrents = computed(() => dashboardStore.selectedTorrents.map(maindataStore.getTorrentByHash).filter(torrent => !!torrent))
+const torrent = computed(() => torrentStore.getTorrentByHash(hash.value))
+const torrents = computed(() => dashboardStore.selectedTorrents.map(torrentStore.getTorrentByHash).filter(torrent => !!torrent))
 const availableCategories = computed(() => [{ name: '' }, ...maindataStore.categories])
 
 async function resumeTorrents() {
-  await maindataStore.resumeTorrents(hashes)
+  await torrentStore.resumeTorrents(hashes)
 }
 
 async function forceResumeTorrents() {
-  await maindataStore.forceResumeTorrents(hashes)
+  await torrentStore.forceResumeTorrents(hashes)
 }
 
 async function pauseTorrents() {
-  await maindataStore.pauseTorrents(hashes)
+  await torrentStore.pauseTorrents(hashes)
 }
 
 function deleteTorrents() {
@@ -60,7 +61,7 @@ function renameTorrents() {
 }
 
 async function forceRecheck() {
-  await maindataStore.recheckTorrents(hashes)
+  await torrentStore.recheckTorrents(hashes)
 }
 
 async function forceReannounce() {
@@ -84,8 +85,8 @@ function hasTag(tag: string) {
 }
 
 async function toggleTag(tag: string) {
-  if (hasTag(tag)) await maindataStore.removeTorrentTags(hashes.value, [tag])
-  else await maindataStore.addTorrentTags(hashes.value, [tag])
+  if (hasTag(tag)) await torrentStore.removeTorrentTags(hashes.value, [tag])
+  else await torrentStore.addTorrentTags(hashes.value, [tag])
 }
 
 async function copyValue(valueToCopy: string) {
@@ -106,7 +107,7 @@ function setShareLimit() {
 
 async function exportTorrents() {
   hashes.value.forEach(hash => {
-    maindataStore.exportTorrent(hash).then(blob => {
+    torrentStore.exportTorrent(hash).then(blob => {
       const url = window.URL.createObjectURL(blob)
       const link = document.createElement('a')
       link.href = url
@@ -170,22 +171,22 @@ const menuData = computed<TRCMenuEntry[]>(() => [
       {
         text: t('dashboard.right_click.priority.top'),
         icon: 'mdi-priority-high',
-        action: async () => await maindataStore.setTorrentPriority(hashes.value, 'topPrio')
+        action: async () => await torrentStore.setTorrentPriority(hashes.value, 'topPrio')
       },
       {
         text: t('dashboard.right_click.priority.increase'),
         icon: 'mdi-arrow-up',
-        action: async () => await maindataStore.setTorrentPriority(hashes.value, 'increasePrio')
+        action: async () => await torrentStore.setTorrentPriority(hashes.value, 'increasePrio')
       },
       {
         text: t('dashboard.right_click.priority.decrease'),
         icon: 'mdi-arrow-down',
-        action: async () => await maindataStore.setTorrentPriority(hashes.value, 'decreasePrio')
+        action: async () => await torrentStore.setTorrentPriority(hashes.value, 'decreasePrio')
       },
       {
         text: t('dashboard.right_click.priority.bottom'),
         icon: 'mdi-priority-low',
-        action: async () => await maindataStore.setTorrentPriority(hashes.value, 'bottomPrio')
+        action: async () => await torrentStore.setTorrentPriority(hashes.value, 'bottomPrio')
       }
     ]
   },
@@ -209,7 +210,7 @@ const menuData = computed<TRCMenuEntry[]>(() => [
     disabledIcon: 'mdi-label-off',
     children: availableCategories.value.map(category => ({
       text: category.name === '' ? t('dashboard.right_click.category.clear') : category.name,
-      action: async () => await maindataStore.setTorrentCategory(hashes.value, category.name)
+      action: async () => await torrentStore.setTorrentCategory(hashes.value, category.name)
     }))
   },
   {
