+++ src/components/Dialogs/AddTorrentDialog.vue
@@ -1,7 +1,7 @@
 <script lang="ts" setup>
 import { useDialog } from '@/composables'
 import { AppPreferences } from '@/constants/qbit'
-import { useAddTorrentStore, useMaindataStore, usePreferenceStore, useVueTorrentStore } from '@/stores'
+import { useAddTorrentStore, useMaindataStore, usePreferenceStore, useTorrentStore, useVueTorrentStore } from '@/stores'
 import { storeToRefs } from 'pinia'
 import { computed, ref } from 'vue'
 import { useI18n } from 'vue-i18n'
@@ -20,6 +20,7 @@ const addTorrentStore = useAddTorrentStore()
 const { urls, files, form } = storeToRefs(addTorrentStore)
 const maindataStore = useMaindataStore()
 const preferenceStore = usePreferenceStore()
+const torrentStore = useTorrentStore()
 const vueTorrentStore = useVueTorrentStore()
 
 const fileOverflowDisplayLimit = 2
@@ -136,7 +137,7 @@ const inactiveSeedingTimeLimit = computed({
 function submit() {
   if (!isFormValid.value) return
 
-  const promise = maindataStore.addTorrents(form.value, files.value)
+  const promise = torrentStore.addTorrents(form.value, files.value)
   .then(() => {
     addTorrentStore.resetForm()
     close()
