+++ src/components/Navbar/SideWidgets/FilterSelect.vue
@@ -1,6 +1,6 @@
 <script lang="ts" setup>
 import { TorrentState } from '@/constants/qbit'
-import { useMaindataStore, useVueTorrentStore } from '@/stores'
+import { useMaindataStore, useTorrentStore, useVueTorrentStore } from '@/stores'
 import { storeToRefs } from 'pinia'
 import { computed } from 'vue'
 import { useI18n } from 'vue-i18n'
@@ -9,12 +9,14 @@ const { t } = useI18n()
 const {
   categories: _categories,
   tags: _tags,
-  trackers: _trackers,
+  trackers: _trackers
+} = storeToRefs(useMaindataStore())
+const {
   statusFilter,
   categoryFilter,
   tagFilter,
   trackerFilter
-} = storeToRefs(useMaindataStore())
+} = storeToRefs(useTorrentStore())
 const vueTorrentStore = useVueTorrentStore()
 
 const statuses = computed(() => Object.values(TorrentState).map(state => (
