+++ src/stores/torrents.ts
@@ -1,0 +1,182 @@
+import { useSearchQuery } from '@/composables'
+import { TorrentState } from '@/constants/qbit'
+import { extractHostname } from '@/helpers'
+import { qbit } from '@/services'
+import { AddTorrentPayload } from '@/types/qbit/payloads'
+import { Torrent } from '@/types/vuetorrent'
+import { defineStore, storeToRefs } from 'pinia'
+import { computed, MaybeRefOrGetter, ref, toValue } from 'vue'
+import { useDashboardStore } from './dashboard'
+
+export const useTorrentStore = defineStore('torrents', () => {
+  const torrents = ref<Torrent[]>([])
+
+  const isTextFilterActive = ref(true)
+  const isStatusFilterActive = ref(true)
+  const isCategoryFilterActive = ref(true)
+  const isTagFilterActive = ref(true)
+  const isTrackerFilterActive = ref(true)
+
+  const textFilter = ref('')
+  const statusFilter = ref<TorrentState[]>([])
+  const categoryFilter = ref<string[]>([])
+  const tagFilter = ref<(string | null)[]>([])
+  const trackerFilter = ref<(string | null)[]>([])
+
+  const torrentsWithFilters = computed(() => {
+    return torrents.value.filter(torrent => {
+      if (statusFilter.value.length > 0 && isStatusFilterActive.value && !statusFilter.value.includes(torrent.state)) return false
+      if (categoryFilter.value.length > 0 && isCategoryFilterActive.value && !categoryFilter.value.includes(torrent.category)) return false
+      if (tagFilter.value.length > 0 && isTagFilterActive.value) {
+        if (torrent.tags.length === 0 && tagFilter.value.includes(null)) return true
+        if (!torrent.tags.some(tag => tagFilter.value.includes(tag))) return false
+      }
+      if (trackerFilter.value.length > 0 && isTrackerFilterActive.value && !trackerFilter.value.includes(extractHostname(torrent.tracker))) return false
+
+      return true
+    })
+  })
+  const filteredTorrents = computed(() => searchQuery.results.value)
+
+  const { sortOptions } = storeToRefs(useDashboardStore())
+  const searchQuery = useSearchQuery(
+    torrentsWithFilters,
+    () => isTextFilterActive.value ? textFilter.value : null,
+    torrent => torrent.name,
+    results => {
+      if (sortOptions.value.isCustomSortEnabled) {
+        if (sortOptions.value.sortBy === 'priority') {
+          results.sort((a, b) => {
+            if (a.priority > 0 && b.priority > 0) return a.priority - b.priority
+            else if (a.priority <= 0 && b.priority <= 0) return a.added_on - b.added_on
+            else if (a.priority <= 0) return 1
+            else return -1
+          })
+        } else {
+          results.sort((a, b) => a[sortOptions.value.sortBy] - b[sortOptions.value.sortBy] || a.added_on - b.added_on)
+        }
+        if (sortOptions.value.reverseOrder) results.reverse()
+      }
+      return results
+    }
+  )
+
+  async function setTorrentCategory(hashes: string[], category: string) {
+    await qbit.setCategory(hashes, category)
+  }
+
+  async function addTorrentTags(hashes: string[], tags: string[]) {
+    await qbit.addTorrentTag(hashes, tags)
+  }
+
+  async function removeTorrentTags(hashes: string[], tags: string[]) {
+    await qbit.removeTorrentTag(hashes, tags)
+  }
+
+  function getTorrentByHash(hash: string) {
+    return torrents.value.find(t => t.hash === hash)
+  }
+
+  function getTorrentIndexByHash(hash: string) {
+    return filteredTorrents.value.findIndex(t => t.hash === hash)
+  }
+
+  async function deleteTorrents(hashes: string[], deleteWithFiles: boolean) {
+    await qbit.deleteTorrents(hashes, deleteWithFiles)
+  }
+
+  async function moveTorrents(hashes: string[], newPath: string) {
+    await qbit.setTorrentLocation(hashes, newPath)
+  }
+
+  async function addTorrents(payload: AddTorrentPayload, torrents: File[]) {
+    return await qbit.addTorrents(payload, torrents)
+  }
+
+  async function getTorrentProperties(hash: string) {
+    return await qbit.getTorrentProperties(hash)
+  }
+
+  async function renameTorrent(hash: string, newName: string) {
+    await qbit.setTorrentName(hash, newName)
+  }
+
+  async function resumeTorrents(hashes: MaybeRefOrGetter<string[]>) {
+    await qbit.resumeTorrents(toValue(hashes))
+  }
+
+  async function forceResumeTorrents(hashes: MaybeRefOrGetter<string[]>) {
+    await qbit.forceStartTorrents(toValue(hashes))
+  }
+
+  async function pauseTorrents(hashes: MaybeRefOrGetter<string[]>) {
+    await qbit.pauseTorrents(toValue(hashes))
+  }
+
+  async function recheckTorrents(hashes: MaybeRefOrGetter<string[]>) {
+    await qbit.recheckTorrents(toValue(hashes))
+  }
+
+  async function setTorrentPriority(hashes: string[], priority: 'increasePrio' | 'decreasePrio' | 'topPrio' | 'bottomPrio') {
+    await qbit.setTorrentPriority(hashes, priority)
+  }
+
+  async function exportTorrent(hash: string) {
+    return await qbit.exportTorrent(hash)
+  }
+
+  return {
+    torrents,
+    isTextFilterActive,
+    isStatusFilterActive,
+    isCategoryFilterActive,
+    isTagFilterActive,
+    isTrackerFilterActive,
+    textFilter,
+    statusFilter,
+    categoryFilter,
+    tagFilter,
+    trackerFilter,
+    torrentsWithFilters,
+    filteredTorrents,
+    searchQuery,
+    setTorrentCategory,
+    addTorrentTags,
+    removeTorrentTags,
+    getTorrentByHash,
+    getTorrentIndexByHash,
+    deleteTorrents,
+    moveTorrents,
+    addTorrents,
+    getTorrentProperties,
+    renameTorrent,
+    resumeTorrents,
+    forceResumeTorrents,
+    pauseTorrents,
+    recheckTorrents,
+    setTorrentPriority,
+    exportTorrent
+  }
+}, {
+  persist: {
+    enabled: true,
+    strategies: [
+      {
+        storage: localStorage,
+        key: 'vuetorrent_torrents',
+        paths: [
+          'isTextFilterActive',
+          'textFilter',
+          'isStatusFilterActive',
+          'statusFilter',
+          'isCategoryFilterActive',
+          'categoryFilter',
+          'isTagFilterActive',
+          'tagFilter',
+          'isTrackerFilterActive',
+          'trackerFilter'
+        ]
+      }
+    ]
+  }
+})
