+++ src/types/qbit/payloads/PeerLogPayload.ts
@@ -1,5 +1,5 @@
 import type { Optional } from '@/types/global'
-import type { BasePayload } from './index'
+import type BasePayload from './BasePayload'
 
 export default interface PeerLogPayload extends BasePayload {
   // Exclude messages with "message id" <= last_known_id (default: -1)
